Added xml functionality
