+++ seir-refactor/scenarioT1_1.py
@@ -6,7 +6,7 @@ r_0 = 1 / 3.5   # rate out of infectious state without testing
 r_R =  1 / 15   # rate out of state hospitalized (H) into recovery or death
 p_V_H = 0.02    # P(Hospitalized | Infected and Vaxxed)
 p_H = 0.2       # P(Hospitalized | Infected)
-p_D = 0.1       # p(Dead | Hospitalized)
+p_D = 0.1       # P(Dead | Hospitalized)
 a_0 = 0.6       # Initial infection rate (proportion/day)
 delta_a = 0.6   # Change in infection rate for a new variant (proportion/day)
 p_e = 0.6       # Transmission rate from a vaccinated person (proportion of unvaccinated rate)
